Create README.md
 *Quran and Sunnah Website

A website dedicated to providing authentic Quranic verses and Sunnah resources to users worldwide.

About the Project

This project aims to spread knowledge online and onsite platform for exploring the Quran and Sunnah. The website will feature:

- Quranic verses with translations and commentary
- Authentic Hadith collections
- Islamic articles and resources
- Prayer timings and Qibla direction

Features

- Responsive design for mobile and desktop devices
- Search functionality for Quranic verses and Hadith
- User-friendly interface for easy navigation
- Islamic calendar integration

Contributing

We welcome contributions to this project! If you'd like to contribute, please


Acknowledgments

We acknowledge the following resources:

- All the four imaams: Hanafi Shafi, Malik and hambali.
- Bukhari, Muslim and others
- inn kathira and others 

Contact

For any questions or feedback, please reach out to [chemicalecho11@gmail.com or Contact +27719679192].
*